+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- scale codded region and viewport size
+- updated to @applitools/driver@1.4.15 (from 1.4.12)
+- updated to @applitools/logger@1.0.11 (from 1.0.10)
+- updated to @applitools/screenshoter@3.3.8 (from 3.3.5)
+- updated to @applitools/snippets@2.1.14 (from 2.1.12)
+- updated to @applitools/types@1.0.25 (from 1.0.24)
+- updated to @applitools/utils@1.2.13 (from 1.2.11)
+
 
 ## 13.0.2 - 2022/2/9